requirements.txt
# Core dependencies
streamlit>=1.31.0
torch>=2.0.0
torchaudio>=2.0.0

# Audio processing and transcription
pyannote.audio>=3.1.0
faster-whisper>=1.0.0

# LangChain for summarization
langchain>=0.1.0
langchain-community>=0.0.20

# Additional dependencies
numpy>=1.24.0
scipy>=1.11.0
